Update MainViewController.java
Updating functionality of methods which are invoking after clicking on playlist, song and song in playlist. Right now they are checking if item was double-clicked. If yes - they are invoking play method on selected song. In case of double clicking on playlist we are invoking play method on first song in selected playlist